+++ processor/src/test/fixtures/input/com/example/prop/ImmutablePropTypeMultiple.java
@@ -15,6 +15,7 @@ import react4j.annotations.Render;
 import react4j.annotations.View;
 
 // Covers multiple immutable prop types as that forces code gen through separate path
+@SuppressWarnings( "React4j:MethodBasedImmutableInput" )
 @View
 abstract class ImmutablePropTypeMultiple
 {