+++ processor/src/test/fixtures/input/com/example/prop/SuppressedMethodImmutableInputWarningModel.java
@@ -8,15 +8,11 @@ import react4j.annotations.View;
 
 @SuppressWarnings( "React4j:MethodBasedImmutableInput" )
 @View
-abstract class ImmutableMixedInputModel
+abstract class SuppressedMethodImmutableInputWarningModel
 {
-  ImmutableMixedInputModel( @Nullable @Input( immutable = true ) final String myImmutableProp )
-  {
-  }
-
   @Nullable
-  @Input
-  abstract String getMyMutableProp();
+  @Input( immutable = true )
+  abstract String getMyProp();
 
   @Nullable
   @Render